Move Attribute classes to Common namespace
