+++ src/SDK/Resource/Detectors/Process.php
@@ -6,7 +6,7 @@ namespace OpenTelemetry\SDK\Resource\Detectors;
 
 use function extension_loaded;
 use function getmypid;
-use OpenTelemetry\SDK\Attributes;
+use OpenTelemetry\SDK\Common\Attribute\Attributes;
 use OpenTelemetry\SDK\Resource\ResourceDetectorInterface;
 use OpenTelemetry\SDK\Resource\ResourceInfo;
 use OpenTelemetry\SemConv\ResourceAttributes;
