+++ src/SDK/Trace/ImmutableSpan.php
@@ -6,7 +6,7 @@ namespace OpenTelemetry\SDK\Trace;
 
 use function max;
 use OpenTelemetry\API\Trace as API;
-use OpenTelemetry\SDK\AttributesInterface;
+use OpenTelemetry\SDK\Common\Attribute\AttributesInterface;
 use OpenTelemetry\SDK\InstrumentationLibrary;
 use OpenTelemetry\SDK\Resource\ResourceInfo;
