+++ src/SDK/Trace/SamplerInterface.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace OpenTelemetry\SDK\Trace;
 
 use OpenTelemetry\Context\Context;
-use OpenTelemetry\SDK\AttributesInterface;
+use OpenTelemetry\SDK\Common\Attribute\AttributesInterface;
 
 /**
  * This interface is used to organize sampling logic.
@@ -22,7 +22,7 @@ interface SamplerInterface
      *                        Typically in situations when the Span to be created starts a new Trace.
      * @param string $spanName Name of the Span to be created.
      * @param int $spanKind Span kind.
-     * @param AttributesInterface|null $attributes Initial set of Attributes for the Span being constructed.
+     * @param \OpenTelemetry\SDK\Common\Attribute\AttributesInterface|null $attributes Initial set of Attributes for the Span being constructed.
      * @param list<LinkInterface> $links Collection of links that will be associated with the Span to be created.
      *                     Typically, useful for batch operations.
      *                     @see https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/overview.md#links-between-spans
