+++ tests/Unit/API/Trace/NonRecordingSpanTest.php
@@ -7,7 +7,7 @@ namespace OpenTelemetry\Tests\API\Unit\Trace;
 use Exception;
 use OpenTelemetry\API\Trace\NonRecordingSpan;
 use OpenTelemetry\API\Trace\SpanContextInterface;
-use OpenTelemetry\SDK\AttributesInterface;
+use OpenTelemetry\SDK\Common\Attribute\AttributesInterface;
 use PHPUnit\Framework\TestCase;
 
 /**
