+++ tests/Unit/Contrib/ZipkinToNewrelicSpanConverterTest.php
@@ -6,7 +6,7 @@ namespace OpenTelemetry\Tests\Unit\Contrib;
 
 use function implode;
 use OpenTelemetry\Contrib\ZipkinToNewrelic\SpanConverter;
-use OpenTelemetry\SDK\Attributes;
+use OpenTelemetry\SDK\Common\Attribute\Attributes;
 use OpenTelemetry\SDK\Resource\ResourceInfo;
 use OpenTelemetry\Tests\Unit\SDK\Util\SpanData;
 use PHPUnit\Framework\TestCase;
